Refactor code in order to help testability and readability
